requirements.txt
djangorestframework==3.9.4
Django==3.1.2
